internal/log: advertise raft on fully qualified domain name
Currently, Raft's address is configured as the transport's local
address, and the server will advertise its address ass ::8400. In order
to use a fully qualified domain name so the node will properly advertise
itself to its cluster and to its client, add a bind address field to the
raft configuration struct.